requirements.txt
# Telegram Bot
aiogram>=3.0

# Video Processing
# Using pre-built wheels for Linux x86_64 (Render platform)
# deepfilternet>=0.3.0
# Note: If this fails to install, comment it out and video processing will be disabled
deepfilternet

moviepy>=1.0.3

# Audio Processing
# librosa>=0.10.0

# MCP Server for Render API
mcp>=1.0.0
httpx>=0.25.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
numpy>=1.24.0
soundfile>=0.12.0
